refactor: organise vault trade test support
