+++ tests/cli/test_vault_test_trade.py
@@ -17,16 +17,16 @@ from tradeexecutor.cli.commands.lagoon_deploy_vault import (
     _write_state_sibling_deployment_artifact,
 )
 from tradeexecutor.cli.commands.vault_test_trade import _validate_vault_test_options
-from tradeexecutor.cli import vault_test_trade_tui as vault_test_trade_tui_module
-from tradeexecutor.cli import (
-    vault_test_trade_simulation as vault_test_trade_simulation_module,
+from tradeexecutor.cli.vault_trade import tui as tui_module
+from tradeexecutor.cli.vault_trade import (
+    simulation as simulation_module,
 )
-from tradeexecutor.cli.vault_test_trade import (
+from tradeexecutor.cli.vault_trade.core import (
     filter_rpc_kwargs_for_vault_specs,
     load_lagoon_deployment,
     parse_vault_ids,
 )
-from tradeexecutor.cli.vault_test_trade_state import (
+from tradeexecutor.cli.vault_trade.state import (
     capture_vault_test_error,
     classify_vault_test_failure,
     create_vault_test_diagnostic_pair,
@@ -36,13 +36,13 @@ from tradeexecutor.cli.vault_test_trade_state import (
     record_attempt_result,
     stamp_position_vault_test_attempt,
 )
-from tradeexecutor.cli.vault_test_trade_tui import (
+from tradeexecutor.cli.vault_trade.tui import (
     VaultChoice,
     VaultSearchScreen,
     VaultTestTradeApp,
 )
-from tradeexecutor.cli.vault_test_trade_setup import load_vault_test_state
-from tradeexecutor.cli.vault_test_trade_simulation import (
+from tradeexecutor.cli.vault_trade.setup import load_vault_test_state
+from tradeexecutor.cli.vault_trade.simulation import (
     SimulatedVaultAttemptTimeout,
     SimulatedVaultRuntime,
     is_simulated_infrastructure_failure,
@@ -50,7 +50,7 @@ from tradeexecutor.cli.vault_test_trade_simulation import (
     raise_simulated_vault_attempt_timeout,
     take_simulated_snapshots,
 )
-from tradeexecutor.cli.vault_test_trade_runner import (
+from tradeexecutor.cli.vault_trade.runner import (
     VaultAttemptContext,
     VaultTestBatchRunner,
     get_bridge_conflict,
@@ -808,7 +808,7 @@ def test_simulated_snapshots_only_touch_attempt_chains(
         return hex(len(snapshotted_connections))
 
     monkeypatch.setattr(
-        vault_test_trade_simulation_module, "make_anvil_custom_rpc_request", snapshot
+        simulation_module, "make_anvil_custom_rpc_request", snapshot
     )
 
     # 2. Take snapshots for an Arbitrum vault with Base as the source.
@@ -1090,17 +1090,17 @@ async def test_vault_main_table_enter_selects_redemption(
     position.simulated = False
     position.is_open.return_value = True
     monkeypatch.setattr(
-        vault_test_trade_tui_module,
+        tui_module,
         "get_latest_vault_position",
         lambda state, vault_spec: position,
     )
     monkeypatch.setattr(
-        vault_test_trade_tui_module,
+        tui_module,
         "get_vault_trade_position",
         lambda state, vault_spec, open_only=False: position,
     )
     monkeypatch.setattr(
-        vault_test_trade_tui_module,
+        tui_module,
         "get_vault_test_status",
         lambda position: "deposited",
     )
