+++ tradeexecutor/cli/vault_trade/__init__.py
@@ -1,0 +1,1 @@
+"""Vault deposit and redemption test-trade support."""
